+++ src/Stormancer.Plugins/Stormancer.Server.TestApp/App.cs
@@ -4,5 +4,9 @@ namespace Stormancer.Server.TestApp
 {
     public class App
     {
+        public void Run(IAppBuilder builder)
+        {
+            builder.AddPlugin(new TestPlugin());
+        }
     }
 }