+++ src/Stormancer.Plugins/Api/Stormancer.Server.Plugins.Api/Changelog.rst
@@ -14,6 +14,7 @@ Unreleased
 Changed
 *******
 - Remove 3 allocations in action call expression generation.
+- 
 Added
 *****
 - Action call generation now maps CancellationToken action parameters to IS2SRequestContext.CancellationToken.