+++ src/Stormancer.Plugins/GameVersion/Stormancer.Server.Plugins.GameVersion/Changelog.rst
@@ -9,6 +9,11 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Uses Users 5.x API
 2.0.2
 -----
 Changed